requirements.txt
fastapi
uvicorn
python-multipart
Jinja2
psycopg2
fastapi[postgres]
python-dotenv
sqlalchemy
psycopg2-binary